Done resize error messages files
